# src/swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "시니나눔",
    "description": "시니나눔 API 문서",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "http",
      "scheme": "bearer",
      "in": "header",
      "bearerFormat": "JWT"
    }
  },
  "paths": {
...
    "/career/": {
      "post": {
        "tags": [
          "Career"
        ],
        "summary": "경력 프로필 생성",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "Career"
        ],
        "summary": "내 경력 프로필 조회",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "tags": [
          "Career"
        ],
        "summary": "경력 프로필 수정",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "profileId": {
                  "type": "number",
                  "example": 111111
                },
                "introduce": {
                  "type": "string",
                  "example": "소개"
                },
                "age": {
                  "type": "string",
                  "example": "60대"
                },
                "field": {
                  "type": "string",
                  "example": "경제"
                },
                "service": {
                  "type": "string",
                  "example": "제공할 서비스"
                },
                "method": {
                  "type": "string",
                  "example": "대면"
                },
                "region": {
                  "type": "string",
                  "example": "종로구"
                },
                "priceType": {
                  "type": "string",
                  "example": "시간당"
                },
                "price": {
                  "type": "number",
                  "example": 1000
                },
                "progressStep": {
                  "type": "number",
                  "example": 8
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/career/list": {
      "get": {
        "tags": [
          "Career"
        ],
        "summary": "경력프로필 목록 불러오기",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/career/{profileId}": {
      "get": {
        "tags": [
          "Career"
        ],
        "summary": "경력프로필 상세조회",
        "description": "",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "   {   introduce: ,   age: 초등,유아,중등,   field: 돌봄,생활,입시,   service: null,   method: 비대면,   region: null,   price: null,   priceType: null,   nickname: 신주현,   gender: 여성,   birthyear: 2001,   profile: http://img1.kakaocdn.net/thumb/R640x640.q70/?fname=http://t1.kakaocdn.net/account_images/default_profile.jpeg,   careerItems: ",
            "introduce": "",
            "age": "초등,유아,중등",
            "field": "돌봄,생활,입시",
            "service": null,
            "method": "비대면",
            "region": null,
            "price": null,
            "priceType": null,
            "nickname": "신주현",
            "gender": "여성",
            "birthyear": "2001",
            "profile": "http://img1.kakaocdn.net/thumb/R640x640.q70/?fname=http://t1.kakaocdn.net/account_images/default_profile.jpeg",
            "careerItems": [],
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/career/filter": {
      "post": {
        "tags": [
          "Career"
        ],
        "summary": "나리 필터링에 맞는 경력프로필 조회",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "method": {
                  "example": "any"
                },
                "priceType": {
                  "example": "any"
                },
                "price": {
                  "example": "any"
                },
                "region": {
                  "example": "any"
                },
                "field": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}